completed migration to next.js and finished build
