+++ app/chapter/[chapterNumber]/page.tsx
@@ -1,11 +1,12 @@
-import { ApolloClient, InMemoryCache, gql } from "@apollo/client";
+import { ApolloClient, gql, InMemoryCache } from "@apollo/client";
+import { getChapterData } from "../../../lib/getChapterData";
 import { Metadata } from "next";
 import ChapterPage from "./chapter-page";
 
 export const metadata: Metadata = {
   title: "Bhagavad Gita App - Chapters",
 };
-export async function getStaticPaths() {
+export async function generateStaticParams() {
   const client = new ApolloClient({
     uri: "https://gql.bhagavadgita.io/graphql",
     cache: new InMemoryCache(),
@@ -24,50 +25,11 @@ export async function getStaticPaths() {
   });
 
   const chapters = data.allGitaChapters?.nodes;
-  const paths = chapters.map(({ chapterNumber }) => {
+  return chapters.map(({ chapterNumber }) => {
     return {
       params: { chapterNumber: chapterNumber.toString() },
     };
   });
-  return {
-    paths,
-    fallback: true,
-  };
-}
-export async function getChapterData(params) {
-  const { chapterNumber } = params;
-  const client = new ApolloClient({
-    uri: "https://gql.bhagavadgita.io/graphql",
-    cache: new InMemoryCache(),
-  });
-
-  const { data } = await client.query({
-    query: gql`
-      query MyQuery {
-        gitaChapterById(id: ${chapterNumber ?? 1}) {
-          chapterSummary
-          chapterNumber
-          nameTranslated
-          versesCount
-          gitaVersesByChapterId {
-            nodes {
-              id
-              verseNumber
-              wordMeanings
-              transliteration
-            }
-          }
-        }
-      }
-    `,
-  });
-
-  const chapterData = data?.gitaChapterById;
-  return {
-    props: {
-      chapterData,
-    },
-  };
 }
 
 export default async function Chapter({ params }) {
