+++ package.json
@@ -25,7 +25,7 @@
     "graphql": "^16.0.1",
     "next": "^13.4.9",
     "next-redux-wrapper": "^7.0.5",
-    "next-themes": "^0.0.15",
+    "next-themes": "^0.2.1",
     "react": "^18.2.0",
     "react-cookie": "^4.1.1",
     "react-dom": "^18.2.0",
